pyproject.toml
# Project metadata
[tool.poetry]
name = "restack-ai-weaviate-curriculum"
version = "0.0.1"
description = "A Restack AI workflow and agent with Weaviate to create a curriculum for a user to learn about a topic"
authors = [
    "Restack Team <service@restack.io>",
]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
watchfiles = "^1.0.0"
pydantic = "^2.10.4"
weaviate-client = "^4.10.4"
google-genai = "0.5.0"

# Build system configuration
restack-ai = "^0.0.52"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# CLI command configuration
[tool.poetry.scripts]
dev = "src.services:watch_services"
services = "src.services:run_services"
schedule = "schedule_workflow:run_schedule_workflow"
interval = "schedule_interval:run_schedule_interval"
calendar = "schedule_calendar:run_schedule_calendar"
